Refactor FX resolution into its own class
fx_muxer is a collection of muxer functionality, framework resolution and command line parsing.
This change moves the framework resolution out into its own fx_resolver.

In the fx_resolver, made some of the methods "this" methods so that they don't have to pass the framework maps in arguments.

No functional changes, just moving code around.
